requirements.txt
Django==1.11.4
Markdown==2.6.9
Pygments==2.2.0
pytz==2017.2
